hang works again
