--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/commandbase/command/drive/DisablePedroTeleOpCommand.java
@@ -1,4 +1,0 @@
-package org.firstinspires.ftc.teamcode.common.commandbase.command.drive;
-
-public class DisablePedroTeleOpCommand {
-}
